+++ apps/backend/src/modules/activities/libs/enums/relation-name.enum.ts
@@ -1,5 +1,7 @@
 const RelationName = {
 	AVATAR_FILE: "avatarFile",
+	COMMENTS: "comments",
+	LIKES: "likes",
 	USER: "user",
 	USER_DETAILS: "userDetails",
 } as const;