+++ packages/shared/src/index.ts
@@ -48,6 +48,7 @@ export {
 } from "./libs/types/types.js";
 export {
 	ActivitiesApiPath,
+	type ActivityCounts,
 	type ActivityCreateRequestDto,
 	type ActivityDeleteRequestDto,
 	type ActivityGetAllResponseDto,
@@ -58,6 +59,10 @@ export {
 	activityCreateFinishSectionValidationSchema,
 	activityDeleteFinishSectionValidationSchema,
 } from "./modules/activities/activities.js";
+export {
+	type ActivityLikeRequestDto,
+	activityLikeChangeValidationSchema,
+} from "./modules/activity-likes/activity-likes.js";
 export { AuthApiPath, AuthError } from "./modules/auth/auth.js";
 export {
 	type ChatMessageCreateRequestDto,
@@ -82,6 +87,19 @@ export {
 	chatCreateValidationSchema,
 	chatIdParameterValidationSchema,
 } from "./modules/chats/chats.js";
+export {
+	type CommentCreateRequestDto,
+	CommentError,
+	type CommentGetAllRequestDto,
+	type CommentGetAllResponseDto,
+	type CommentUpdateRequestDto,
+	type CommentWithRelationsResponseDto,
+	CommentsApiPath,
+	commentCreateBodyValidationSchema,
+	commentGetAllQueryValidationSchema,
+	commentIdParameterValidationSchema,
+	commentTextValidationSchema,
+} from "./modules/comments/comments.js";
 export {
 	type CourseSectionAddRequestDto,
 	type CourseSectionDto,
@@ -158,6 +176,7 @@ export {
 } from "./modules/user-notifications/user-notifications.js";
 export {
 	type UserAuthResponseDto,
+	type UserDetailsResponseDto,
 	UserError,
 	type UserGetAllResponseDto,
 	type UserGetByIdRequestDto,