+++ packages/shared/src/modules/comments/libs/types/comment-response-dto.type.ts
@@ -1,0 +1,8 @@
+type CommentResponseDto = {
+	activityId: number;
+	id: number;
+	text: string;
+	userId: number;
+};
+
+export { type CommentResponseDto };